refactor: Avatar feature moving to own component
Avatar feature separated from profile into
its own component which to keep code
clean and organised
