+++ dream-big-ui/src/app/Components/profile/profile.component.scss
@@ -31,7 +31,7 @@
   body, input[text] {
     color: #333;
   }
-  
+
   button {
     border: none;
     cursor: pointer;
@@ -39,132 +39,60 @@
     font-size: 1rem;
     padding: 1rem;
   }
-  
+
   input[type="text"] {
     box-sizing: border-box;
     width: 100%;
     padding: .5rem;
   }
-  
+
   .container {
     display: flex;
     justify-content: center;
     flex-wrap: wrap;
   }
-  
-  .container.avatar {
-    vertical-align: top;
-    text-align: center;
-  }
-  
   .box {
     background-color: #eeedf2;
     padding: 1rem;
     margin: 0.5rem 2rem;
   }
-  
+
   .box:first-child {
     margin-top: 0;
   }
-  
+
   .box > div {
     display: flex;
     flex-direction: column;
   }
-  
+
   .flex {
     display: flex;
     justify-content: space-between;
     flex-direction: row;
   }
-  
+
   .flex:not(:last-child) {
     margin-bottom: 1rem;
   }
-  
+
   .flex > *:not(:last-child) {
     margin-right: 2rem;
   }
-  
-  .random {
-    background-image: url('data:image/svg+xml;utf8,<?xml version="1.0" encoding="utf-8"?><svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 122.88 105.71" style="enable-background:new 0 0 122.88 105.71" xml:space="preserve"><style type="text/css">.st0{fill-rule:evenodd;clip-rule:evenodd;}</style><g><path class="st0" d="M0,79.45c-0.02-1.95,0.76-3.06,2.51-3.18h14.08c5.98,0,8.89,0.16,13.98-3.91c1.08-0.86,2.1-1.86,3.06-3 c4.55-5.41,6.17-11.96,7.87-18.9C44.79,37,50.03,22.78,63.98,17.15c7.94-3.2,18.82-2.59,27.41-2.59h5.27l0.01-10.05 c0-5.01,1.18-5.88,4.79-2.45l19.55,18.58c2.36,2.24,2.03,3.7-0.22,5.86L101.49,45c-3.37,3.41-4.89,2.45-4.82-2.26v-11.8 c-34-0.52-32.57,1.67-42.05,34.09c-3.5,10.04-8.81,17.08-15.59,21.69c-7.09,4.82-13.68,6.39-22.02,6.39H6.65 C0.71,93.11,0,92.83,0,86.75V79.45L0,79.45z M0.23,26.26c-0.02,1.95,0.76,3.06,2.51,3.18h14.7c5.98,0,8.89-0.16,13.98,3.91 c1.08,0.86,2.1,1.86,3.06,3c1.16,1.38,2.13,2.84,2.96,4.35c1.5-4.69,3.36-9.29,5.82-13.5c0.7-1.19,1.44-2.35,2.23-3.48 c-1.74-1.8-3.61-3.37-5.61-4.73c-7.09-4.82-13.68-6.39-22.02-6.39H6.88c-5.94,0-6.65,0.28-6.65,6.36V26.26L0.23,26.26z M53.57,80.45c2.96,3.42,6.63,6.24,11.27,8.11c7.94,3.2,18.21,2.59,26.8,2.59h5.27l0.01,10.05c0,5.01,1.18,5.88,4.79,2.45 l19.55-18.58c2.36-2.24,2.03-3.7-0.22-5.86l-19.3-18.5c-3.37-3.41-4.89-2.45-4.82,2.26v11.8c-24.78,0.38-30.42-0.69-35.32-13.84 c-0.27,0.94-0.64,2.23-1.93,6.65c-0.03,0.1-0.06,0.19-0.09,0.28l0,0C57.91,72.62,55.9,76.79,53.57,80.45L53.57,80.45z"/></g></svg>');
-    background-repeat: no-repeat;
-    background-position: center;
-    width: 100%;
-    height: 2rem;
-    margin-bottom: 0.5rem;
-  }
-  
-  .random.small {
-    height: 1.8rem;
-    min-width: 1.8rem;
-    margin-bottom: 0;
-  }
-  
+
   button {
     border-radius: 0;
     background: #fff;
     vertical-align: bottom;
     margin: 0 0.7rem 0.7rem 0;
   }
-  
+
   button:last-child {
     margin-right: 0;
   }
-  
+
   button:hover {
     color: #333;
     background: #DDDDEE;
   }
-  
-  input[type="color"] {
-     height: 3.5rem;
-     width: 3.5rem;
-    margin: 0 0.7rem 0.7rem 0;
-    cursor: pointer;
-  }
-  
-  .iteration {
-    display: inline-block;
-    font-size: 1rem;
-    padding: 0.1rem;
-    border: 1px solid #444;
-    width: auto;
-    margin-bottom: 0.3rem;
-  }
-  
-  .iterate {
-    font-weight: bold;
-    font-size: 1.5rem;
-  }
-  
-  button.download {
-    border: 2px solid #333;
-    margin-left: 2rem;
-    font-weight: bold;
-  }
-  button.download > span, button.download > svg {
-    display: inline-block;
-    vertical-align: middle;
-  }
-  
-  button.download > svg {
-    width: 2rem;
-    margin-right: 0.5rem;
-  }
-  .avatar {
-    position: sticky;
-    top: 0;
-    background: #fff;
-    width: 100%;
-    text-align: center;
-  }
-  
-  @media (min-width: 755px) {
-    .avatar {
-      width: auto;
-      align-self: flex-start;
-    }
-  }
-  
-  
+
